refactor frontend testing utilities
