+++ src/web/frontend/__init__.py
@@ -41,8 +41,6 @@ from .frontend_testing import (
     FrontendTestRunner,
     TestResult,
     TestSuite,
-    TestUtilities,
-    MockDataGenerator,
 )
 
 # Version information
@@ -73,8 +71,6 @@ __all__ = [
     "FrontendTestRunner",
     "TestResult",
     "TestSuite",
-    "TestUtilities",
-    "MockDataGenerator",
 ]
