+++ src/web/frontend/testing/ui_utils.py
@@ -1,0 +1,54 @@
+"""Utilities for simulating frontend UI interactions in tests."""
+from datetime import datetime
+
+from ..frontend_app import UIComponent, create_component
+from ..frontend_router import RouteConfig, NavigationState
+
+
+class UIInteractionUtilities:
+    """Create mock frontend entities for testing UI behaviour."""
+
+    def create_mock_component(
+        self, component_type: str = "TestComponent"
+    ) -> UIComponent:
+        """Return a mock :class:`UIComponent` instance."""
+        return create_component(
+            component_type,
+            {
+                "id": "test-id",
+                "className": "test-class",
+                "data-testid": "test-component",
+            },
+        )
+
+    def create_mock_route(self, path: str = "/test") -> RouteConfig:
+        """Return a mock :class:`RouteConfig` instance."""
+        return RouteConfig(
+            path=path,
+            name="test-route",
+            component="TestComponent",
+            props={"title": "Test Page"},
+            meta={"requiresAuth": False},
+            children=[],
+            guards=[],
+            middleware=[],
+            lazy_load=False,
+            cache=True,
+            created_at=datetime.now(),
+            updated_at=datetime.now(),
+        )
+
+    def create_mock_navigation_state(self) -> NavigationState:
+        """Return a mock :class:`NavigationState` instance."""
+        return NavigationState(
+            current_route="/test",
+            previous_route="/",
+            route_params={"id": "123"},
+            query_params={"page": "1"},
+            navigation_history=["/", "/test"],
+            breadcrumbs=[
+                {"path": "/", "name": "Home"},
+                {"path": "/test", "name": "Test"},
+            ],
+            timestamp=datetime.now(),
+        )
